Remove NuGet package dependencies from Datadog.Trace and Datadog.Trace.ClrProfiler.Managed (#761)
- Remove `Microsoft.CSharp` NuGet package reference. This was only necessary for using the `dynamic` keyword, so those have been refactored to objects that use our existing `ObjectExtensions` methods
- Remove `System.Runtime.InteropServices.RuntimeInformation` NuGet package reference. This is responsible for the additional `System.Runtime.InteropServices.RuntimeInformation.dll` on all platforms. It is built-in for `netstandard2.0` so the logic that uses the new API's was ifdef'd out of the .NET Framework build targets.
- Remove `Microsoft.Win32.Registry` NuGet package reference. This was no longer needed after refactoring of `FrameworkDescription` code.
