+++ deploy/Datadog.Trace.ClrProfiler.WindowsInstaller/Files.Managed.Net45.GAC.wxs
@@ -45,11 +45,6 @@
               Source="$(var.TracerHomeDirectory)\net45\System.Diagnostics.DiagnosticSource.dll"
               KeyPath="yes" Checksum="yes" Assembly=".net"/>
       </Component>
-      <Component Win64="$(var.Win64)">
-        <File Id="net45_GAC_System.Runtime.InteropServices.RuntimeInformation.dll"
-              Source="$(var.TracerHomeDirectory)\net45\System.Runtime.InteropServices.RuntimeInformation.dll"
-              KeyPath="yes" Checksum="yes" Assembly=".net"/>
-      </Component>
     </ComponentGroup>
   </Fragment>
 </Wix>
