+++ src/Datadog.Trace/Logging/DatadogLogging.cs
@@ -13,8 +13,6 @@ namespace Datadog.Trace.Logging
 {
     internal static class DatadogLogging
     {
-        private const string NixDefaultDirectory = "/var/log/datadog/dotnet";
-
         private static readonly long? MaxLogFileSize = 10 * 1024 * 1024;
         private static readonly LoggingLevelSwitch LoggingLevelSwitch = new LoggingLevelSwitch(LogEventLevel.Information);
         private static readonly ILogger SharedLogger = null;
@@ -40,7 +38,15 @@ namespace Datadog.Trace.Logging
                     MaxLogFileSize = maxLogSize;
                 }
 
-                var logDirectory = GetLogDirectory();
+                string logDirectory = null;
+                try
+                {
+                    logDirectory = GetLogDirectory();
+                }
+                catch
+                {
+                    // Do nothing when an exception is thrown for attempting to access the filesystem
+                }
 
                 // ReSharper disable once ConditionIsAlwaysTrueOrFalse
                 if (logDirectory == null)
@@ -141,41 +147,34 @@ namespace Datadog.Trace.Logging
             //   - Path.GetTempPath
             if (logDirectory == null)
             {
+#if NETFRAMEWORK
+                logDirectory = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"Datadog .NET Tracer", "logs");
+#else
                 if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
                 {
-                    var windowsDefaultDirectory = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"Datadog .NET Tracer", "logs");
-                    if (Directory.Exists(windowsDefaultDirectory))
-                    {
-                        logDirectory = windowsDefaultDirectory;
-                    }
+                    logDirectory = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"Datadog .NET Tracer", "logs");
                 }
                 else
                 {
                     // Linux
-                    if (Directory.Exists(NixDefaultDirectory))
-                    {
-                        logDirectory = NixDefaultDirectory;
-                    }
-                    else
-                    {
-                        try
-                        {
-                            var di = Directory.CreateDirectory(NixDefaultDirectory);
-                            logDirectory = NixDefaultDirectory;
-                        }
-                        catch
-                        {
-                            // Unable to create the directory meaning that the user
-                            // will have to create it on their own.
-                        }
-                    }
+                    logDirectory = "/var/log/datadog/dotnet";
                 }
+#endif
             }
 
-            if (logDirectory == null)
+            if (!Directory.Exists(logDirectory))
             {
-                // Last effort at writing logs
-                logDirectory = Path.GetTempPath();
+                try
+                {
+                    Directory.CreateDirectory(logDirectory);
+                }
+                catch
+                {
+                    // Unable to create the directory meaning that the user
+                    // will have to create it on their own.
+                    // Last effort at writing logs
+                    logDirectory = Path.GetTempPath();
+                }
             }
 
             return logDirectory;
